+++ sequencer/Sequence/sequence.py
@@ -1,721 +1,15 @@
+
+from sequencer.Sequence.event import Event, EventBehavior, Jump, Ramp, Digital, RampType
+from sequencer.Sequence.channel import Channel, Analog_Channel, Digital_Channel
+from sequencer.Sequence.time_frame import TimeInstance
+
+from typing import List, Optional, Callable
 import copy
-import numpy as np
-from abc import ABC, abstractmethod
-from enum import Enum
-from typing import Callable, List, Optional, Union,Tuple, Dict, Any 
-import bisect
-import json 
-import os
-import sys
-import time
-import copy
+import json
+from sequencer.Sequence.utils import tuple_to_string,string_to_tuple
 import matplotlib.pyplot as plt
-import itertools
 
 
-
-from sympy import symbols
-from sympy.parsing.sympy_parser import parse_expr
-from sympy.utilities.lambdify import lambdify
-
-
-def tuple_to_string(tuple_input):
-    """
-    Convert a tuple of three strings to one string.
-    
-    Args:
-    tuple_input (tuple): A tuple containing three strings.
-    
-    Returns:
-    str: A single string combining the three input strings.
-    """
-    if len(tuple_input) != 3:
-        raise ValueError("Input tuple must contain exactly three strings.")
-    
-    return "|".join(tuple_input)
-
-def string_to_tuple(string_input):
-    """
-    Convert one string to a tuple of three strings.
-    
-    Args:
-    string_input (str): A string containing three parts separated by '|'.
-    
-    Returns:
-    tuple: A tuple containing three strings.
-    """
-    parts = string_input.split("|")
-    
-    if len(parts) != 3:
-        raise ValueError("Input string must contain exactly two '|' separators.")
-    
-    return tuple(parts)
-
-
-
-def exp_to_func(exp_str):
-    x = symbols('t')
-    expr = parse_expr(exp_str)
-    return lambdify(x, expr)
-def find_min_max(func, start, end, step=0.000001):
-    """
-    Find the minimum and maximum values of a function within a given range using NumPy.
-    
-    Args:
-    func (callable): The function to evaluate
-    start (float): The start of the range
-    end (float): The end of the range
-    step (float): The step size for the range (default is 1)
-    
-    Returns:
-    tuple: (min_value, max_value, t_min, t_max)
-    """
-    t_range = np.arange(start, end + step, step)
-    values = func(t_range)
-    
-    max_value = np.max(values)
-    min_value = np.min(values)
-    
-    t_max = t_range[np.argmax(values)]
-    t_min = t_range[np.argmin(values)]
-    
-    return min_value, max_value, t_min, t_max
-
-
-class Parameter:
-    def __init__(self, name: str,event: 'Event',parameter_origin):
-        self.name = name
-        self.event = event
-        self.parameter_origin = parameter_origin
-    def get_value(self): 
-        return self.event.get_event_attributes()[self.parameter_origin]
-
-        
-        
-
-
-
-class RampType(Enum):
-    LINEAR = 'Linear'
-    QUADRATIC = 'Quadratic'
-    EXPONENTIAL = 'Exponential'
-    LOGARITHMIC = 'Logarithmic'
-    GENERIC = 'Generic'
-    MINIMUM_JERK = 'Minimum Jerk'
-
-    def __str__(self):
-        return self.value
-
-
-
-class EventBehavior(ABC):
-    @abstractmethod
-    def get_value_at_time(self, t: float) -> float:
-        pass
-
-class Jump(EventBehavior):
-    def __init__(self, target_value: float):
-        self.target_value = target_value
-
-    def edit_jump(self, target_value: float):
-        self.target_value = target_value
-
-
-    def get_value_at_time(self, t: float) -> float:
-        return self.target_value
-    
-    def __repr__(self) -> str:
-        return f"Jump({self.target_value})"
-
-
-class Ramp(EventBehavior):
-    def __init__(self, start_time_instance:'TimeInstance',end_time_instance:'TimeInstance', ramp_type: RampType = RampType.LINEAR, start_value: float = 0, end_value: float = 1, func_text: str = None, resolution=0.001):
-        # if start_value == end_value:
-        #     raise ValueError("start_value and end_value must be different")
-        
-        if end_time_instance =="temp":
-            pass 
-        elif start_time_instance.get_absolute_time() >= end_time_instance.get_absolute_time():
-            raise ValueError("start_time_instance must be less than end_time_instance")
-        
-        
-        self.start_time_instance = start_time_instance
-        self.end_time_instance = end_time_instance
-
-        
-        if ramp_type == RampType.EXPONENTIAL and (start_value == 0 or end_value == 0):
-            raise ValueError("For exponential ramp, start_value and end_value must be non-zero")
-
-        if resolution < 0.000001:
-            raise ValueError("resolution must be at least 0.000001")
-        
-        self.ramp_type = ramp_type
-        self.start_value = start_value
-        self.end_value = end_value
-        self.resolution = resolution
-        self.func_text = func_text
-        if self.ramp_type == RampType.GENERIC:
-            if func_text:
-                self.func = exp_to_func(func_text)
-            else:
-                raise ValueError("func can only be provided for generic ramp type, not for other ramp types")
-        else:
-            self._set_func()
-    
-    def get_duration(self):
-        return self.end_time_instance.get_absolute_time() - self.start_time_instance.get_absolute_time()
-
-    def _set_func(self):
-        if self.end_time_instance == "temp":
-            return
-        
-        duration = self.get_duration()
-        if self.ramp_type == RampType.LINEAR:
-            self.func = lambda t: self.start_value + (self.end_value - self.start_value) * (t / duration)
-        elif self.ramp_type == RampType.QUADRATIC:
-            self.func = lambda t: self.start_value + (self.end_value - self.start_value) * (t / duration)**2
-        elif self.ramp_type == RampType.EXPONENTIAL:
-            self.func = lambda t: self.start_value * (self.end_value / self.start_value) ** (t / duration)
-        elif self.ramp_type == RampType.LOGARITHMIC:
-            self.func = lambda t: self.start_value + (self.end_value - self.start_value) * (np.log10(t + 1) / np.log10(duration + 1))
-        elif self.ramp_type == RampType.MINIMUM_JERK:
-            self.func = lambda t: self.start_value + (self.end_value - self.start_value) * (10 * (t/duration)**3 - 15 * (t/duration)**4 + 6 * (t/duration)**5)
-        else:
-            raise ValueError("Invalid ramp type")
-    def get_start_value(self):
-        return self.func(0)
-
-    def get_end_value(self):
-        return self.func(self.get_duration())
-    def edit_ramp(self, 
-                        start_time_instance: Optional['TimeInstance'] = None,
-                        end_time_instance: Optional['TimeInstance'] = None, 
-                        ramp_type: Optional[RampType] = None,
-                        start_value: Optional[float] = None,
-                        end_value: Optional[float] = None,
-                        func_text: str = None,
-                        resolution: Optional[float] = None):
-        
-        new_start_time_instance = start_time_instance if start_time_instance is not None else self.start_time_instance
-        new_end_time_instance = end_time_instance if end_time_instance is not None else self.end_time_instance
-
-        if ramp_type is not None:
-            try:
-                new_ramp_type = ramp_type
-            except KeyError:
-                # Handle the case where the ramp_type string is not valid
-                raise ValueError(f"Invalid ramp type: {ramp_type}")
-        else:
-            new_ramp_type = self.ramp_type
-
-        new_start_value = start_value if start_value is not None else self.start_value
-        new_end_value = end_value if end_value is not None else self.end_value
-        new_resolution = resolution if resolution is not None else self.resolution
-        
-        # if new_start_value == new_end_value:
-        #     raise ValueError("start_value and end_value must be different")
-        
-        if new_start_time_instance.get_absolute_time() >= new_end_time_instance.get_absolute_time():
-            raise ValueError("duration must be negative")
-        
-        if new_ramp_type == RampType.EXPONENTIAL and (new_start_value == 0 or new_end_value == 0):
-            raise ValueError("For exponential ramp, start_value and end_value must be non-zero")
-        
-        if new_resolution < 0.000001:
-            raise ValueError("resolution must be at least 0.000001")
-        
-        # Apply changes only after validation
-        self.start_time_instance = new_start_time_instance
-        self.end_time_instance = new_end_time_instance
-        self.ramp_type = new_ramp_type
-        self.start_value = new_start_value
-        self.end_value = new_end_value
-        self.resolution = new_resolution
-
-        if func_text:
-            self.func_text = func_text
-            self.func = exp_to_func(func_text)
-        else:
-            self._set_func()
-
-
-    def get_value_at_time(self, t: float) -> float:
-        if 0 <= t <= self.get_duration():
-            return self.func(t)
-        else:
-            return self.func(self.get_duration())
-    
-    def __repr__(self) -> str:
-        return f"Ramp({self.get_duration()}, {self.ramp_type.value}, {self.get_start_value()}, {self.get_end_value()})"
-class Digital(EventBehavior):
-    def __init__(self, target_value: float):
-        if target_value not in [0, 1]:
-            raise ValueError("target_value must be 0 or 1")
-        self.target_value = target_value
-
-    def edit_digital(self, target_value: float):
-        if target_value not in [0, 1]:
-            raise ValueError("target_value must be 0 or 1")
-        self.target_value = target_value
-
-    def get_value_at_time(self, t: float) -> float:
-        return self.target_value
-    
-    def __repr__(self) -> str:
-        return f"Digital({self.target_value})"
-class Channel:
-    def __init__(self, name: str, card_number: int, channel_number: int, reset: bool, reset_value: float):
-        self.name = name
-        self.card_number = card_number
-        self.channel_number = channel_number
-        self.reset = reset
-        self.reset_value = reset_value
-        self.events: List[Event] = []
-    
-    def add_event(self, event: 'Event'):
-        index = bisect.bisect_left([e.get_start_time() for e in self.events], event.get_start_time())
-        self.events.insert(index, event)
-    def sort_events(self):
-        self.events.sort(key=lambda event: event.get_start_time())
-
-    def __repr__(self) -> str:
-        return (
-            f"Channel(\n"
-            f"   name={self.name},\n"
-            f"   card_number={self.card_number},\n"
-            f"   channel_number={self.channel_number},\n"
-            f"   reset={self.reset},\n"
-            f"   reset_value={self.reset_value},\n"
-            f"   events={self.events}\n"
-            f")"
-        )
-    
-    #vituak function to get the channel attributes
-    def get_channel_attributes(self):
-        return {
-            "name": self.name,
-            "card_number": self.card_number,
-            "channel_number": self.channel_number,
-            "reset": self.reset,
-            "reset_value": self.reset_value
-        }
-
-    def check_for_overlapping_events(self):
-        self.events.sort(key=lambda event: event.get_start_time())
-        for i in range(len(self.events) - 1):
-            current_event = self.events[i]
-            next_event = self.events[i + 1]
-            if current_event.get_end_time() > next_event.get_start_time():
-                raise ValueError(f"Events {current_event} and {next_event} on channel {self.name} overlap.")
-
-    def detect_a_ramp(self,time_instance:'TimeInstance'):
-        # check if the time instance between two time instances has a ramp
-        # get all the ramps in the channel: 
-        for event in time_instance.events:
-            if event.channel == self:
-                return None
-        for event in time_instance.ending_ramps:
-            if event.channel == self:
-                return None
-        ramps = [event for event in self.events if isinstance(event.behavior, Ramp)] 
-        for ramp in ramps:
-            if ramp.start_time_instance.get_absolute_time() < time_instance.get_absolute_time() < ramp.end_time_instance.get_absolute_time():
-                value = ramp.behavior.get_value_at_time(time_instance.get_absolute_time() - ramp.start_time_instance.get_absolute_time())
-                return ramp, value
-        return None
-    
-    def get_event_by_time_instance(self,time_instance:'TimeInstance'):
-        for event in self.events:
-            if event.start_time_instance == time_instance:
-                return (event,"start")
-        for event in self.events:
-            if event.end_time_instance == time_instance:
-                return (event,"end")
-        return None
-    
-
-             
-    def __eq__(self, other: object) -> bool:
-        return self.name == other.name and self.card_number == other.card_number and self.channel_number == other.channel_number and self.reset == other.reset and self.reset_value == other.reset_value
-
-class Analog_Channel(Channel):
-    def __init__(self, name: str, card_number: int, channel_number: int, reset: bool = False, reset_value: float = 0, default_voltage_func: Callable[[float], float] = lambda a: a, max_voltage: float = 10, min_voltage: float = -10,LIMIT=65535, RANGE=20, OFFSET=10):
-        super().__init__(name, card_number, channel_number, reset, reset_value)
-        self.default_voltage_func = default_voltage_func  # Should map from whatever value to a value between -10 to 10 
-        self.max_voltage = max_voltage
-        self.min_voltage = min_voltage
-        self.LIMIT=LIMIT
-        self.RANGE=RANGE
-        self.OFFSET=OFFSET
-
-    def discretize(self,val) : # Should map from -10:10 to 0:65535
-        return (val + self.OFFSET)/self.RANGE * self.LIMIT
-  
-
-    def __repr__(self) -> str:
-        return (
-            f"Analog_Channel(\n"
-            f"   name={self.name},\n"
-            f"   card_number={self.card_number},\n"
-            f"   channel_number={self.channel_number},\n"
-            f"   reset={self.reset},\n"
-            f"   reset_value={self.reset_value},\n"
-            f"   max_voltage={self.max_voltage},\n"
-            f"   min_voltage={self.min_voltage},\n"
-            f"   events={self.events}\n"
-            f")"
-        )
-    def get_channel_attributes(self):
-        return {
-            "type": "analog",
-            "name": self.name,
-            "card_number": self.card_number,
-            "channel_number": self.channel_number,
-            "reset": self.reset,
-            "reset_value": self.reset_value,
-            "max_voltage": self.max_voltage,
-            "min_voltage": self.min_voltage,
-            "LIMIT":self.LIMIT,
-            "RANGE":self.RANGE,
-            "OFFSET":self.OFFSET,
-            # "default_voltage_func": self.default_voltage_func
-
-        }
-
-    
-    def __eq__(self, other: object) -> bool:
-        return super().__eq__(other) and self.default_voltage_func == other.default_voltage_func and self.max_voltage == other.max_voltage and self.min_voltage == other.min_voltage
-
-
-class Digital_Channel(Channel):
-    def __init__(self, name: str, card_number: int, channel_number: int,  reset: bool = False, reset_value: float = 0):
-        super().__init__(name, card_number, channel_number, reset, reset_value)
-
-    def __eq__(self, other: object) -> bool:
-        return super().__eq__(other) 
-
-    def __repr__(self) -> str:
-        return (
-            f"Digital_Channel(\n"
-            f"   name={self.name},\n"
-            f"   card_number={self.card_number},\n"
-            f"   channel_number={self.channel_number},\n"
-            f"   reset={self.reset},\n"
-            f"   reset_value={self.reset_value},\n"
-            f"   events={self.events}\n"
-            f")"
-        )
-    def get_channel_attributes(self):
-        return {
-            "type": "digital",
-            "name": self.name,
-            "card_number": self.card_number,
-            "channel_number": self.channel_number,
-            "reset": self.reset,
-            "reset_value": self.reset_value,
-        }
-    
-class Event:
-    def __init__(self, channel: Optional[Union[Digital_Channel, Analog_Channel]] , behavior: EventBehavior,comment:str ="",start_time_instance: Optional['TimeInstance'] = None,end_time_instance: Optional['TimeInstance'] = None):
-
-
-        self.channel = channel
-        self.behavior = behavior
-        self.comment= comment
-        self.is_sweept = False
-        self.reference_original_event = self
-        self.associated_parameters = []
-        # check if the start_time_instance and channels already have events, raise an error if there is already an event at the same time and channel
-        for event in start_time_instance.events:
-            if event.channel == channel:
-                raise ValueError(f"Event already exists on channel {channel.name} at time instance {start_time_instance.name}")
-        if end_time_instance and end_time_instance != "temp":
-            for event in end_time_instance.events:
-                if event.channel == channel:
-                    raise ValueError(f"Event already exists on channel {channel.name} at time instance {end_time_instance.name}")
-            
-        if start_time_instance is not None:
-            self.start_time_instance = start_time_instance
-        else:
-            raise ValueError("start_time_instance is required")
-        
-        if isinstance(behavior, Ramp):
-            if end_time_instance=="temp":
-                pass
-            else:
-                if end_time_instance is None:
-                    raise ValueError("end_time_instance is required for Ramp events")
-                else :
-                    self.end_time_instance = end_time_instance
-        else:
-            self.end_time_instance = self.start_time_instance
-
-            
-        
-        
-
-
-        if isinstance(behavior, Ramp) and not isinstance(behavior, Jump):
-            
-            self.end_time_instance =  end_time_instance 
-        else:
-            self.end_time_instance = self.start_time_instance
-
-        if end_time_instance != "temp":
-            self.check_for_overlap(channel, behavior, self.get_start_time(), self.get_end_time())
-            self.check_if_generic_ramp_is_valid()
-        
-
-
-        self.children: List[Event] = []
-        index = bisect.bisect_left([e.get_start_time() for e in self.channel.events], self.get_start_time())
-        self.channel.events.insert(index, self)
-        self.start_time_instance.add_event(self)
-        if self.end_time_instance != "temp":
-            if isinstance(behavior, Ramp) and not isinstance(behavior, Jump) and self.end_time_instance != self.start_time_instance:
-                self.end_time_instance.add_ending_ramp(self)
-    
-    def check_if_generic_ramp_is_valid(self):
-        if isinstance(self.behavior, Ramp) and self.behavior.ramp_type == RampType.GENERIC:
-            # get the min and max values of the function
-            min_value, max_value, _, _ = find_min_max(self.behavior.func, 0, self.behavior.get_duration())
-            if min_value < self.channel.min_voltage or max_value > self.channel.max_voltage:
-                raise ValueError(f"Generic ramp values are out of range for channel {self.channel.name} with min voltage {self.channel.min_voltage} and max voltage {self.channel.max_voltage}")
-        else:
-            return
-            
-    def get_start_time(self):
-        return self.start_time_instance.get_absolute_time()
-
-    def get_end_time(self):
-        return self.end_time_instance.get_absolute_time()
-    
-    
-
-    def get_event_attributes(self):
-        if isinstance(self.behavior, Jump):
-            
-                
-            return {
-                "type": "jump",
-                "jump_target_value": self.behavior.target_value,
-                "start_time": self.get_start_time(),
-                "channel_name": self.channel.name,
-                "comment":self.comment,
-                "start_time_instance":self.start_time_instance.name,
-                "is_sweept":self.is_sweept
-                }
-
-        elif isinstance(self.behavior, Ramp):
-            return {
-                "type": "ramp",
-                "ramp_type": self.behavior.ramp_type.value,
-                "start_value": self.behavior.start_value,
-                "end_value": self.behavior.end_value,
-                "func": self.behavior.func_text,
-                "resolution": self.behavior.resolution,
-                "channel_name": self.channel.name,
-                "comment":self.comment,
-                "start_time_instance":self.start_time_instance.name,
-                "end_time_instance":self.end_time_instance.name,
-                "is_sweept":self.is_sweept
-                }   
-        elif isinstance(self.behavior, Digital):
-            return {
-                "type": "digital",
-                "target_value": self.behavior.target_value,
-                "start_time": self.get_start_time(),
-                "channel_name": self.channel.name,
-                "comment":self.comment,
-                "start_time_instance":self.start_time_instance.name,
-                "is_sweept":self.is_sweept,
-                }
-
-
-
-    def check_for_overlap(self, channel: Channel, behavior: EventBehavior, start_time: float, end_time: float):
-        for event in channel.events:
-            if not (end_time < event.get_start_time() or start_time > event.get_end_time()):
-                if isinstance(behavior, Jump) and isinstance(event.behavior, Jump):
-                    if start_time == event.get_start_time():
-                        raise ValueError(f"Cannot have more than one jump at the same time on channel {channel.name}.")
-                elif isinstance(behavior, Jump) and isinstance(event.behavior, Ramp):
-                    if start_time != event.get_end_time():
-                        raise ValueError(f"Jump events can only be added at the end of a ramp on channel {channel.name}.")
-                else:
-                    raise ValueError(f"Events on channel {channel.name} overlap with existing event {event.behavior} from {event.get_start_time()} to {event.get_end_time()}.")
-
-class TimeInstance:
-    def __init__(self, name: str, parent: Optional['TimeInstance'] = None, relative_time: int = 0):
-        self.name: str = name
-        self.parent: Optional['TimeInstance'] = parent
-        self.relative_time: int = relative_time
-        self.children: List['TimeInstance'] = []
-        self.events: List[Event] = []
-        self.ending_ramps: List[Event] = []
-        self.is_sweept = False
-        if parent:
-            parent.children.append(self)
-        else :
-            # check if the relative time is not 0 for the root time instance
-            if relative_time != 0:
-                raise ValueError("Relative time must be 0 for the root time instance")
-            
-    def add_event(self, event: 'Event') -> None:
-        self.events.append(event)
-    def add_ending_ramp(self, event: 'Event') -> None:
-        self.ending_ramps.append(event)
-
-    def get_child_time_instance_by_name(self, name: str) -> Optional['TimeInstance']:
-        for child in self.children:
-            if child.name == name:
-                return child
-        return
-    def get_descendant_by_name(self, name: str) -> Optional['TimeInstance']:
-        for child in self.children:
-            if child.name == name:
-                return child
-            descendant = child.get_descendant_by_name(name)
-            if descendant:
-                return descendant
-        return None
-    
-    
-    def edit_parent(self, new_parent_name: str) -> None:
-        # check if the new parent is not the current time instance
-        if self.name == new_parent_name:
-            raise ValueError("Cannot set the parent to self")
-        # check if the parent is not a descendant of the current time instance
-        new_parent = self.get_time_instance_by_name(new_parent_name)
-        
-        if self == new_parent:
-            raise ValueError("Cannot set the parent to self")
-        # check if the parent is not a descendant of the current time instance
-        if new_parent in self.get_all_children():
-            raise ValueError("Cannot set the parent to a descendant")
-        # check the parent belongs to the same sequence
-        if self.get_root() != new_parent.get_root():
-            raise ValueError("Cannot set the parent to a time instance from a different sequence")
-        # check if editing the parent will result in negative absolute time
-        if new_parent.get_absolute_time() + self.relative_time < 0: 
-            raise ValueError("new relative time will result in negative absolute time")
-        
-        if self.parent:
-            self.parent.children.remove(self)
-        self.parent = new_parent
-        new_parent.children.append(self)
-
-    def edit_name(self, new_name: str) -> None:
-        
-        root = self.get_root()
-        all_children = root.get_all_children()
-        
-        for child in all_children:
-            if child.name == new_name:
-                raise Exception("Name already exists")
-        
-        self.name = new_name
-
-    def edit_relative_time(self, new_relative_time: int) -> None:
-        # check if the new relative time will result in negative absolute time
-        if self.parent is not None:
-            if self.parent.get_absolute_time() + new_relative_time < 0:
-                raise ValueError("new relative time will result in negative absolute time")
-        self.relative_time = new_relative_time
-
-    def get_absolute_time(self) -> int:
-        if self.parent is None:
-            return self.relative_time
-        return self.relative_time + self.parent.get_absolute_time()
-
-    def add_child_time_instance(self, name: str, relative_time: int) -> 'TimeInstance':
-        #  check if the relative time will result in negative absolute time
-        if self.get_absolute_time() + relative_time < 0:
-            raise ValueError("new relative time will result in negative absolute time")
-        return TimeInstance(name, parent=self, relative_time=relative_time)
-
-    def delete_self(self) -> None:
-        
-        if self.parent is None:
-            raise Exception("Cannot delete root time frame")
-
-        # check if deleting self will result in negative absolute time for any of the children
-
-        # get all children of the current time instance
-        children = self.get_all_children()
-        delta = - self.relative_time 
-        for child in children:
-            if child.get_absolute_time() + delta < 0:
-                raise ValueError("Deleting this time instance will result in negative absolute time for one of the children")
-            
-        for child in self.children:
-            child.parent = self.parent
-            self.parent.children.append(child)
-        self.parent.children.remove(self)
-        self.children = []
-
-    def get_root(self) -> 'TimeInstance':
-        if self.parent is None:
-            return self
-        return self.parent.get_root()
-
-    def get_all_children(self) -> List['TimeInstance']:
-        children: List['TimeInstance'] = []
-        for child in self.children:
-            children.append(child)
-            children += child.get_all_children()
-        return children
-    def get_all_time_instances(self) -> List['TimeInstance']:
-        time_instances: List['TimeInstance'] = []
-        time_instances.append(self)
-        time_instances += self.get_all_children()
-        return time_instances
-    
-    def get_all_time_instances_after_me(self) -> List['TimeInstance']:
-        # get root time instance
-        root = self.get_root()
-        # get all time instances in the sequence
-        all_time_instances = root.get_all_time_instances()
-        sort_time_instances = sorted(all_time_instances, key=lambda x: x.get_absolute_time())
-        index = sort_time_instances.index(self)
-        return sort_time_instances[index + 1:]
-    
-
-    
-    def get_time_instance_by_name(self, name: str) -> Optional['TimeInstance']:
-        if self.name == name:
-            return self
-        # get all children of the current time instance
-        children = self.get_root().get_all_children()
-        for child in children:
-            if child.name == name:
-                return child
-        return None
-    
-
-    def print_children(self, depth: int = 0) -> None:
-        print("  " * depth + str(self))
-        for child in self.children:
-            child.print_children(depth + 1)
-
-    def get_events_string(self, depth: int = 0) -> str:
-        events_string: str = ""
-        for event in self.events:
-            events_string += "  " * depth + str(event) + "\n"
-        for child in self.children:
-            events_string += child.get_events_string(depth + 1)
-        return events_string
-
-    def create_a_deep_copy_of_all_frames(self) -> 'TimeInstance':
-        root = self.get_root()
-        new_root = copy.deepcopy(root)
-        return new_root
-
-    def __str__(self) -> str:
-        return f"TimeInstance(name={self.name}, absolute_time={self.get_absolute_time()}, events={self.events}, relative_time={self.relative_time})"
-
-    def __repr__(self) -> str:
-        return self.__str__()
 
 
 class Sequence:
@@ -1067,6 +361,16 @@ class Sequence:
         for channel in temp_sequence.channels:
             channel.check_for_overlapping_events()
         # apply the changes to the original sequence
+        # unstack sweep parameters in all the events
+        for event in events:
+            if event.is_sweept:
+                event.is_sweept = False
+                self.unstack_sweep_parameter(event)
+        
+        # unstack the sweep values of the time instance
+        if time_instance.is_sweept:
+            self.unstack_sweep_parameter(time_instance)
+
         time_instance = self.root_time_instance.get_time_instance_by_name(name)
         events = time_instance.events
         for event in events:
