+++ temp_seq.json
@@ -1,0 +1,3 @@
+{
+    "0": "{\n    \"name\": \"M\",\n    \"sweep_dict\": {},\n    \"sweep_values\": [],\n    \"channels\": [\n        {\n            \"type\": \"analog\",\n            \"name\": \"A1\",\n            \"card_number\": 1,\n            \"channel_number\": 1,\n            \"reset\": false,\n            \"reset_value\": 0.0,\n            \"max_voltage\": 10.0,\n            \"min_voltage\": -10.0,\n            \"LIMIT\": 65535,\n            \"RANGE\": 20,\n            \"OFFSET\": 10\n        }\n    ],\n    \"root_time_instance\": {\n        \"relative_time\": 0,\n        \"is_sweept\": false,\n        \"name\": \"root\",\n        \"events\": [],\n        \"ending_ramps\": [],\n        \"children\": [\n            {\n                \"relative_time\": 1,\n                \"is_sweept\": false,\n                \"name\": \"t1\",\n                \"events\": [],\n                \"ending_ramps\": [],\n                \"children\": []\n            },\n            {\n                \"relative_time\": 2,\n                \"is_sweept\": false,\n                \"name\": \"t2\",\n                \"events\": [],\n                \"ending_ramps\": [],\n                \"children\": [\n                    {\n                        \"relative_time\": 6,\n                        \"is_sweept\": false,\n                        \"name\": \"t5\",\n                        \"events\": [],\n                        \"ending_ramps\": [],\n                        \"children\": []\n                    }\n                ]\n            },\n            {\n                \"relative_time\": 4,\n                \"is_sweept\": false,\n                \"name\": \"t3\",\n                \"events\": [],\n                \"ending_ramps\": [],\n                \"children\": []\n            }\n        ]\n    }\n}"
+}
